move controllers to viewer component
