+++ src/components/viewer/WebGLPointsController.ts
@@ -1,11 +1,8 @@
 import { mat3 } from "gl-matrix";
 
-import markersUrl from "../assets/markers/markers.png?url";
-import pointsFragmentShader from "../assets/shaders/points.frag?raw";
-import pointsVertexShader from "../assets/shaders/points.vert?raw";
-import { PointsData } from "../data/points";
-import { TableData } from "../data/table";
-import { Layer } from "../model/layer";
+import { PointsData } from "../../data/points";
+import { TableData } from "../../data/table";
+import { Layer } from "../../model/layer";
 import {
   DEFAULT_POINT_COLOR,
   DEFAULT_POINT_COLORS,
@@ -21,7 +18,7 @@ import {
   PointsLayerConfig,
   RawPointsLayerConfig,
   createPointsLayerConfig,
-} from "../model/points";
+} from "../../model/points";
 import {
   Color,
   DrawOptions,
@@ -30,11 +27,14 @@ import {
   Rect,
   isTableGroupsColumn,
   isTableValuesColumn,
-} from "../types";
-import ColorUtils from "../utils/ColorUtils";
-import HashUtils from "../utils/HashUtils";
-import TransformUtils from "../utils/TransformUtils";
-import WebGLUtils from "../utils/WebGLUtils";
+} from "../../types";
+import ColorUtils from "../../utils/ColorUtils";
+import HashUtils from "../../utils/HashUtils";
+import TransformUtils from "../../utils/TransformUtils";
+import WebGLUtils from "../../utils/WebGLUtils";
+import markersUrl from "../assets/markers/markers.png?url";
+import pointsFragmentShader from "../assets/shaders/points.frag?raw";
+import pointsVertexShader from "../assets/shaders/points.vert?raw";
 import WebGLControllerBase from "./WebGLControllerBase";
 
 export default class WebGLPointsController extends WebGLControllerBase {
